GithubAction/app/src/main/java/com/sava/githubaction/AppMethods.java
package com.sava.githubaction;

import java.nio.ByteBuffer;
import java.text.SimpleDateFormat;

public class AppMethods {

    public static long convertHexToInt(String hexString) {
        return Long.parseLong(hexString, 16);
    }

    public static int convertByteArrayToInt(byte[] bytes) {
        ByteBuffer wrapped = ByteBuffer.wrap(bytes); // big-endian by default
        return wrapped.getShort();
    }

    private static final char[] HEX_ARRAY = "0123456789ABCDEF".toCharArray();

    public static String bytesToHex(byte[] bytes) {
        char[] hexChars = new char[bytes.length * 2];
        for (int j = 0; j < bytes.length; j++) {
            int v = bytes[j] & 0xFF;
            hexChars[j * 2] = HEX_ARRAY[v >>> 4];
            hexChars[j * 2 + 1] = HEX_ARRAY[v & 0x0F];
        }
        return new String(hexChars);
    }

    public static String getCurrentTime() {
        SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS");
        return formatter.format(new java.util.Date());
    }
}
